refactor: split world.ts into 5 domain files + barrel (#478)
Moves transport, vehicles, tools, space, and weather (+ ADVANCED_WEATHER_*)
into worldData/ subfiles. world.ts becomes a 5-line barrel — zero breaking
changes for consumers.
